ADD: ex01, in toFloat() i was losing the fractional part after the devison bc fixed-point is an int as well for 256 with the cast we make sure that we save everything that we need

+++ ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: tndreka < tndreka@student.42heilbronn.d    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/04 13:21:56 by tndreka           #+#    #+#             */
-/*   Updated: 2025/07/06 02:40:56 by tndreka          ###   ########.fr       */
+/*   Updated: 2025/07/06 02:46:40 by tndreka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,7 +97,7 @@ float Fixed::toFloat(void)const
 {
 	float to_f;
 
-	to_f = fixed_point / 256;
+	to_f = (float)fixed_point / 256;
 	return to_f;
 }
